scheduler/requirements.txt
confluent_kafka==2.2.0
cryptography==41.0.1
jsonschema==4.18.4
ruamel.yaml==0.17.32
schedule==1.2.0
python-dateutil==2.8.2